requirements.txt
cryptography>=41.0.0
Pillow>=10.0.0
pyzbar>=0.1.9
pyotp>=2.8.0 